Include electric bikes into GUI
This commit handles display of electric bikes by adding an 'ebikes'
area to the 'free bikes' and 'empty slots' legacy areas of the GUI.
By default this additional area is non visible, and it shows up only
if the bike system supports it. In this case, the 'free bikes' area
is turned into a 'regular bikes' area : it displays the number of
non-electric bikes only, and the icon is replaced by a green one.

This update is done on 4 pieces of GUI's code :
- StationActivity.java which inflates activity_station.xml
- StationsListAdapter.java which inflates station_list_item.xml
- MapActivity.java which inflates bonuspack_bubble.xml
- StationsListAppWidgetService.java which inflates app_widget_item.xml